+++ go-backend/model/xray_client.go
@@ -14,8 +14,6 @@ type XrayClient struct {
 	ExpTime        *int64 `gorm:"column:exp_time" json:"expTime"`
 	LimitIp        int    `gorm:"column:limit_ip;default:0" json:"limitIp"`
 	Reset          int    `gorm:"column:reset;default:0" json:"reset"`
-	TgId           string `gorm:"column:tg_id" json:"tgId"`
-	SubId          string `gorm:"column:sub_id" json:"subId"`
 	Enable         int    `gorm:"column:enable" json:"enable"`
 	Remark         string `gorm:"column:remark" json:"remark"`
 	CreatedTime    int64  `gorm:"column:created_time" json:"createdTime"`